refactor(dyslexia): extract article styles, expand selector coverage and add persistence tests
- Extracted dyslexia styles from ArticleScreen to utils/dyslexiaStyles

- Improved article style specificity by including ul, ol, blockquote

- Added useDyslexiaMode unit tests for optimistic update fallback
